# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ITPSM API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/score-evaluations": {
            "get": {
                "tags": [
                    "Assign evaluation score"
                ],
                "summary": "Get evaluation score list",
                "description": "Returns evaluation score list and allows to filter by student id or evaluation id ",
                "operationId": "getScore-evaluations",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "filter format like:{\"query\":[{\"field\":\"se.student_id\",\"op\":\"=\",\"data\":\"2\"}]}",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assign evaluation score"
                ],
                "summary": "Assign evaluation score",
                "description": "Assign evaluation score",
                "operationId": "postScore-evaluations",
                "parameters": [
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Evaluation score",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student to assign score",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "evaluation_id",
                        "in": "query",
                        "description": "Evaluation to assign score",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/score-evaluations/{id}": {
            "put": {
                "tags": [
                    "Assign evaluation score"
                ],
                "summary": "Update evaluation score",
                "description": "Update evaluation score",
                "operationId": "putScore-evaluations",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Score evaluation id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "Evaluation score",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Student to assign score",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "evaluation_id",
                        "in": "query",
                        "description": "Evaluation to assign score",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assign evaluation score"
                ],
                "summary": "Delete score-evaluations by id",
                "description": "Delete score-evaluations by id",
                "operationId": "Delete score-evaluations by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "score-evaluations id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/score/insertGrades": {
            "post": {
                "tags": [
                    "Assign evaluation score"
                ],
                "summary": "Save evaluation score",
                "description": "Save evaluation score",
                "operationId": "postGradesEvaluations",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}